requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.6
Django==3.2.7
django-sslserver 
idna==3.2
numpy==1.21.2
Pillow==8.3.2
pytz==2021.1
requests==2.26.0
sqlparse==0.4.2
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.9.1+cpu
torchvision==0.10.1+cpu
typing-extensions==3.10.0.2
urllib3==1.26.6
